+++ src/pages/OrderFinish/hooks/useCancelPayment.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { cancelPayment } from '@/api/payments';
 import { CancelPaymentRequest } from '@/api/payments/entity';
-import { backButtonTapped } from '@/util/ts/bridge';
+import { backButtonTapped } from '@/util/bridge/nativeAction';
 
 export default function useCancelPayment(paymentId: number) {
   return useMutation({
